+++ packages/nuxt-docs/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sit-onyx/nuxt-docs
 
+## 1.0.0-beta.43
+
+### Patch Changes
+
+- dfe66db: fix bug where markdown files are not loaded when using `@nuxt/content >= 3.5.0`
+
+## 1.0.0-beta.42
+
+### Patch Changes
+
+- Updated dependencies [1a4890b]
+  - sit-onyx@1.0.0-beta.229
+  - @sit-onyx/nuxt@1.0.0-beta.231
+
+## 1.0.0-beta.41
+
+### Patch Changes
+
+- Updated dependencies [dc832e4]
+  - sit-onyx@1.0.0-beta.228
+  - @sit-onyx/nuxt@1.0.0-beta.230
+
+## 1.0.0-beta.40
+
+### Patch Changes
+
+- Updated dependencies [d149c34]
+  - sit-onyx@1.0.0-beta.227
+  - @sit-onyx/nuxt@1.0.0-beta.229
+
 ## 1.0.0-beta.39
 
 ### Patch Changes