+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,23 @@
 # sit-onyx
 
+## 1.0.0-beta.229
+
+### Minor Changes
+
+- 1a4890b: feat: implement new `OnyxFileUpload` component
+
+## 1.0.0-beta.228
+
+### Patch Changes
+
+- dc832e4: fix: prevent hydration errors for OnyxFlyout and OnyxTooltip
+
+## 1.0.0-beta.227
+
+### Patch Changes
+
+- d149c34: fix(OnyxNavBar): fix runtime error when OnyxNavItems were mounted asynchronously
+
 ## 1.0.0-beta.226
 
 ### Patch Changes